Add 1.24 support and remove all plugin changes
Upstream AWS QuickStart template has changed to install the EBS CSI
plugin.  It's likely we won't have to do this ourselves now.  Remove
everything related to this, and get rid of some configuration we were
setting ("do not add an OIDC to the cluster") which is now required and
has been removed upstream.