Fix eks-prerequisites Parameters and Lambda runtime
The problem with the Lambda was quite interesting/annoying - this
depends on some Layers setup in eks-prerequisites.  We are not pinning
these layers because they are built/packaged Lambdas, so we take the
approach of copying these to the XRd S3 bucket in publish-s3-bucket, and
we do not pin to a specific version.

The Layer had been updated to use Python 3.8, whereas our Lambda was
using 3.7.  This manifested itself in a very cryptic error saying that
"aws could not be found" - i.e. sounds like the Layer itself hadn't been
set up properly, but the actual root cause was that the shebang in the
"aws" script (which was present and was on the PATH) referenced 3.8
which did not exist.