Simplify publish-s3-bucket syncing
* Do not error if the bucket already exists
* Sync the QS bucket functions/packages to the XRd bucket rather than
  manually copying specific artefacts.  Note this means we copy all
artefacts rather than the select few.  This is deliberate - before we
were copying only what was necessary for amazon-eks-functions to work,
and relying on amazon-eks-prerequisites to set up a temporary lambda
bucket with the rest of the packaged lambdas.  Now we point
amazon-eks-prerequisites at the XRd bucket and just copy everything from
the XRd bucket to the lambda bucket.