WIP merge different platform tests
Now that we have a test for persistence, which applies to both
ControlPlane and vRouter singleton applications, it doesn't make sense
to test these charts separately.

Instead plan is to test the Overlay and Singleton applications
separately, parametrizing the Singleton application tests by
ControlPlane and vRouter repository and tag args to pytest.